% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columbus.R
\docType{data}
\name{columbus}
\alias{columbus}
\title{Columbus Crime 1980}
\format{
An sf data frame with 49 rows, 20 variables, and a geometry column:
\describe{
\item{	AREA	}{	neighborhood area (computed by ArcView)	}
\item{	PERIMETER	}{	neighborhood perimeter (computed by ArcView)	}
\item{	COLUMBUS	}{	internal polygon ID (generated by ArcView)	}
\item{	COLUMBUS_I	}{	internal polygon ID (geneated by ArcView)	}
\item{	POLYID	}{	neighborhood ID, used in GeoDa User’s Guide and tutorials	}
\item{	NEIG	}{	neighborhood ID, used in Spatial Econometrics examples	}
\item{	HOVAL	}{	housing value (in $1,000)	}
\item{	INC	}{	household income (in $1,000)	}
\item{	CRIME	}{	residential burglaries and vehicle thefts per 1000 households	}
\item{	OPEN	}{	open space (area)	}
\item{	PLUMB	}{	percent housing units without plumbing	}
\item{	DISCBD	}{	distance to CBD	}
\item{	X	}{	centroid x coordinate (in arbitrary digitizing units)	}
\item{	Y	}{	centroid y coordinate (in arbitrary digitizing units)	}
\item{	NSA	}{	north-south indicator variable (North = 1)	}
\item{	NSB	}{	other north-south indicator variable (North = 1)	}
\item{	EW	}{	east-west indicator variable (East = 1)	}
\item{	CP	}{	core-periphery indicator variable (Core = 1)	}
\item{	THOUS	}{	constant (= 1000)	}
\item{	NEIGNO	}{	another neighborhood ID variable (NEIG + 1000)	}
}
}
\source{
Anselin, Luc (1988). Spatial Econometrics. Boston, Kluwer Academic, Table 12.1, p. 189.
}
\usage{
columbus
}
\description{
Crime, housing and income data for 49 neighborhoods in Columbus, OH, 1980. Textbook example.
}
\details{
Sf object, unprojected. EPSG 4326: WGS84.
}
\examples{
if (requireNamespace("sf", quietly = TRUE)) {
  library(sf)
  data(columbus)

  plot(columbus["CRIME"])
}
}
\keyword{datasets}
