% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_pop.R
\docType{data}
\name{community_pop}
\alias{community_pop}
\title{Chicago Population Change 2010}
\format{
A data frame with 77 rows and 2 variables:
\describe{
\item{	NID	}{	Neighborhood ID	}
\item{	POP2010	}{ Population 2010}
}
}
\source{
Official City of Chicago Website \url{https://www.chicago.gov/city/en/depts/dcd/supp_info/community_area_2000and2010censuspopulationcomparisons.html}
}
\usage{
community_pop
}
\description{
Population in Chicago community areas 2010.
}
\details{
Non-spatial object
}
\keyword{datasets}
