% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elections1216.R
\docType{data}
\name{elections1216}
\alias{elections1216}
\title{2012 and 2016 Presidential Elections}
\format{
An sf data frame with 3108 rows, 74 variables, and a geometry column:
\describe{
\item{	STATEFP	}{	State FIPS code	}
\item{	COUNTYFP	}{	County FIPS code	}
\item{	GEOID	}{	Geographic area ID	}
\item{	ALAND	}{	Land area	}
\item{	AWATER	}{	Water area	}
\item{	area_name	}{	County name	}
\item{	state_abbr	}{	State abbreviation	}
\item{	PST045214	}{	Population, 2014 estimate	}
\item{	PST040210	}{	Population, 2010 (April 1) estimates base	}
\item{	PST120214	}{	Population, percent change - April 1, 2010 to July 1, 2014	}
\item{	POP010210	}{	Population, 2010	}
\item{	AGE135214	}{	Persons under 5 years, percent, 2014	}
\item{	AGE295214	}{	Persons under 18 years, percent, 2014	}
\item{	AGE775214	}{	Persons 65 years and over, percent, 2014	}
\item{	SEX255214	}{	Female persons, percent, 2014	}
\item{	RHI125214	}{	White alone, percent, 2014	}
\item{	RHI225214	}{	Black or African American alone, percent, 2014	}
\item{	RHI325214	}{	American Indian and Alaska Native alone, percent, 2014	}
\item{	RHI425214	}{	Asian alone, percent, 2014	}
\item{	RHI525214	}{	Native Hawaiian and Other Pacific Islander alone, percent, 2014	}
\item{	RHI625214	}{	Two or More Races, percent, 2014	}
\item{	RHI725214	}{	Hispanic or Latino, percent, 2014	}
\item{	RHI825214	}{	White alone, not Hispanic or Latino, percent, 2014	}
\item{	POP715213	}{	Living in same house 1 year & over, percent, 2009-2013	}
\item{	POP645213	}{	Foreign born persons, percent, 2009-2013	}
\item{	POP815213	}{	Language other than English spoken at home, pct age 5+, 2009-2013	}
\item{	EDU635213	}{	High school graduate or higher, percent of persons age 25+, 2009-2013	}
\item{	EDU685213	}{	Bachelor’s degree or higher, percent of persons age 25+, 2009-2013	}
\item{	VET605213	}{	Veterans, 2009-2013	}
\item{	LFE305213	}{	Mean travel time to work (minutes), workers age 16+, 2009-2013	}
\item{	HSG010214	}{	Housing units, 2014	}
\item{	HSG445213	}{	Homeownership rate, 2009-2013	}
\item{	HSG096213	}{	Housing units in multi-unit structures, percent, 2009-2013	}
\item{	HSG495213	}{	Median value of owner-occupied housing units, 2009-2013	}
\item{	HSD410213	}{	Households, 2009-2013	}
\item{	HSD310213	}{	Persons per household, 2009-2013	}
\item{	INC910213	}{	Per capita money income in past 12 months (2013 dollars), 2009-2013	}
\item{	INC110213	}{	Median household income, 2009-2013	}
\item{	PVY020213	}{	Persons below poverty level, percent, 2009-2013	}
\item{	BZA010213	}{	Private nonfarm establishments, 2013	}
\item{	BZA110213	}{	Private nonfarm employment, 2013	}
\item{	BZA115213	}{	Private nonfarm employment, percent change, 2012-2013	}
\item{	NES010213	}{	Nonemployer establishments, 2013	}
\item{	SBO001207	}{	Total number of firms, 2007	}
\item{	SBO315207	}{	Black-owned firms, percent, 2007	}
\item{	SBO115207	}{	American Indian- and Alaska Native-owned firms, percent, 2007	}
\item{	SBO215207	}{	Asian-owned firms, percent, 2007	}
\item{	SBO515207	}{	Native Hawaiian- and Other Pacific Islander-owned firms, percent, 2007	}
\item{	SBO415207	}{	Hispanic-owned firms, percent, 2007	}
\item{	SBO015207	}{	Women-owned firms, percent, 2007	}
\item{	MAN450207	}{	Manufacturers shipments, 2007 ($1,000)	}
\item{	WTN220207	}{	Merchant wholesaler sales, 2007 ($1,000)	}
\item{	RTN130207	}{	Retail sales, 2007 ($1,000)	}
\item{	RTN131207	}{	Retail sales per capita, 2007	}
\item{	AFN120207	}{	Accommodation and food services sales, 2007 ($1,000)	}
\item{	BPS030214	}{	Building permits, 2014	}
\item{	LND110210	}{	Land area in square miles, 2010	}
\item{	POP060210	}{	Population per square mile, 2010	}
\item{	Demvotes16	}{	Votes for Democratic candidate in 2016 presidential election	}
\item{	GOPvotes16	}{	Votes for Republican candidate in 2016 presidential election	}
\item{	total_2016	}{	Total number of votes cast in 2016 presidential election	}
\item{	pct_dem_16	}{	Votes for Democratic candidate as percent of total votes	}
\item{	pct_gop_16	}{	Votes for Republican candidate as percent of total votes	}
\item{	diff_2016	}{	Difference between the number of votes for Republican and Democratic candidates	}
\item{	pct_pt_16	}{	The number in diff_2016 expressed as a percent of the total votes. Negative if fewer votes were cast for the Democratic candidate	}
\item{	total_2012	}{	Total number of votes cast in 2012 presidential election	}
\item{	Demvotes12	}{	Votes for Democratic candidate as percent of total votes	}
\item{	GOPvotes12	}{	Votes for Republican candidate as percent of total votes	}
\item{	pct_dem_12	}{	Votes for Democratic candidate as percent of total votes	}
\item{	pct_gop_12	}{	Votes for Republican candidate as percent of total votes	}
\item{	diff_2012	}{	Difference between the number of votes for Republican and Democratic candidates	}
\item{	pct_pt_12	}{	The number in diff_2012 expressed as a percent of the total votes. Negative if fewer votes were cast for the Democratic candidate	}
}
}
\source{
\url{https://github.com/tonmcg/US_County_Level_Election_Results_08-16} and \url{https://www.kaggle.com/benhamner/2016-us-election}
}
\usage{
elections1216
}
\description{
US Presidential Election results in 2012 and 2016, by county.
}
\details{
Sf object, unprojected. EPSG 102434: WGS84.
}
\examples{
if (requireNamespace("sf", quietly = TRUE)) {
  library(sf)
  data(elections1216)

  plot(elections1216["diff_2012"])
}
}
\keyword{datasets}
