% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodaData-package.R
\docType{package}
\name{geodaData-package}
\alias{geodaData}
\alias{geodaData-package}
\title{geodaData: Spatial Analysis Datasets for Teaching}
\description{
Stores small spatial datasets used to teach basic spatial analysis
    concepts. Datasets are based off of the 'GeoDa' software workbook and data
    site <https://geodacenter.github.io/data-and-lab/> developed by Luc Anselin
    and team at the University of Chicago. Datasets are stored as 'sf' objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/spatialanalysis/geodaData}
  \item Report bugs at \url{https://github.com/spatialanalysis/geodaData/issues}
}

}
\author{
\strong{Maintainer}: Angela Li \email{ali6@uchicago.edu} (\href{https://orcid.org/0000-0002-8956-419X}{ORCID})

Authors:
\itemize{
  \item Karina Acosta \email{acostak@uchicago.edu} (\href{https://orcid.org/0000-0001-5103-1147X}{ORCID}) [contributor]
}

Other contributors:
\itemize{
  \item Luc Anselin (Creator of original spatial datasets) [contributor]
}

}
\keyword{internal}
