% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile.R
\docType{data}
\name{juvenile}
\alias{juvenile}
\title{Juvenile Offenders in Cardiff, UK}
\format{
An sf data frame with 168 rows, 4 variables, and a geometry column:
\describe{
\item{	ID	}{	location identifier	}
\item{	X	}{	X-coordinate	}
\item{	Y	}{	Y-coordinate	}
}
}
\source{
Bailey, T. and A. Gatrell (1995). Interactive Spatial Data Analysis. New York: Wiley, p. 95.
}
\usage{
juvenile
}
\description{
Residences of juvenile offenders in Cardiff, UK.
}
\details{
Sf object, undefined Coordinate Reference System.
}
\examples{
if (requireNamespace("sf", quietly = TRUE)) {
  library(sf)
  data(juvenile)

  plot(juvenile["Y"])
}
}
\keyword{datasets}
