% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kchomesale.R
\docType{data}
\name{kchomesale}
\alias{kchomesale}
\title{2014-15 Home Sales in King County, WA}
\format{
List of two sf dataframes. Use \code{list2env(kchomesale, .GlobalEnv)} to unlist in working environment. \strong{kchomesales_points} includes homesales points with with 21613 rows, 21 variables, and a geometry column and \strong{kc_zones} includes polygons of zones for King County (85 rows, 7 variables and a geometry column).
\describe{
\item{	date	}{	date sold	}
\item{	price	}{	sale price	}
\item{	bedrooms	}{	number of bedrooms	}
\item{	bathrooms	}{	number of bathrooms	}
\item{	sqft__ving	}{	size of living area in square feet	}
\item{	sqft_lot	}{	size of the lot in square feet	}
\item{	floors	}{	number of floors	}
\item{	waterfront	}{	‘1’ if the property has a waterfront, ‘0’ if not.	}
\item{	view	}{	An index from 0 to 4 of how good the view of the property was (definition:)	}
\item{	condition	}{	condition of the house, ranked from 1 to 5	}
\item{	grade	}{	Classification by construction quality which refers to the types of materials used and the quality of workmanship. Buildings of better quality (higher grade) cost more to build per unit of measure and command higher value. (definition:)	}
\item{	sqft_above	}{	square feet above ground	}
\item{	sqft__ment	}{	square feet below ground	}
\item{	yr_built	}{	year built	}
\item{	yr_re_ated	}{	year renovated. ‘0’ if never renovated	}
\item{	zipcode	}{	5 digit zip code	}
\item{	lat, long	}{	latitude and longitude	}
\item{	squft__ng15	}{	average size of closest 15 houses, in square feet	}
\item{	OBJECTID	}{	Polygon ID	}
\item{	ZIP	}{	ZIP code as int }
\item{	ZIPCODE	}{	ZIP code as chr }
\item{	COUNTY	}{	County code }
\item{	Shape_leng	}{	Polygon length in meters  }
\item{	Shape_Area	}{	Polygon area in meters }
}
}
\source{
King County Home sales obtained from \url{https://www.kaggle.com/harlfoxem/housesalesprediction/discussion} and King county polygons from \url{https://gis-kingcounty.opendata.arcgis.com/datasets/zipcodes-for-king-county-and-surrounding-area-shorelines-zipcode-shore-area}
}
\usage{
kchomesale
}
\description{
Home sales prices and charactiersitcs for Seattle and King County, WA (May 2014 - 2015).
}
\details{
List of two Sf objects, unprojected. EPSG 4326: WGS84.
}
\examples{
if (requireNamespace("sf", quietly = TRUE)) {
  library(sf)
  data(kchomesale)

  plot(kchomesale[[1]]["price"])
}
}
\keyword{datasets}
