% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasrosas2.R
\docType{data}
\name{lasrosas2}
\alias{lasrosas2}
\title{Yield and Fertilizer in Las Rosas farm (Argentina) (2001)}
\format{
An sf data frame with 1705 rows, 27 variables, and a geometry column:
\describe{
\item{	ID (01)	}{	identifier	}
\item{	TOP2	}{	topography dummy, Slope E	}
\item{	TOP3	}{	topography dummy, Hilltop	}
\item{	TOP4	}{	topography dummy, Slope W	}
\item{	NXTOP2 TO NXTOP4	}{	interaction Nitrogen - topography zone	}
\item{	LONGITUDE	}{	longitude	}
\item{	LATITUDE	}{	latitude	}
\item{	OBS	}{	observation number	}
\item{	YIELD	}{	corn yield (quintals per hectare)	}
\item{	N	}{	Nitrogen fertilizer application (kg per hectare)	}
\item{	N2	}{	Nitrogen squared	}
\item{	TOPO	}{	zone: Low E (1), Slope E (2), Hilltop (3), Slope W (4)	}
\item{	BV	}{	brightness value (proxy for low organic matter content)	}
\item{	BV2	}{	brightness squared	}
\item{	NXBV	}{	interaction Nitrogen - brightness	}
\item{	BVXT2 TO BVXT4	}{	interaction brightness topographic zone	}
\item{	BV2XT2 TO BV2XT4	}{	interaction squared brightness topographic zone	}
\item{	SAT (99)	}{	Red to NIR ratio (proxy for low organic matter content)	}
\item{	SAT2 (99)	}{	Sat squared	}
\item{	NXSAT (99)	}{	interaction Nitrogen - Red to NIR ratio	}
\item{	SATXT2 TO SATXT4 (99)	}{	interaction Sat - topographic zone	}
\item{	SAT2XT2 TO SAT2XT4 (99)	}{	interaction Sat squared - topographic zone	}
}
}
\source{
Rodolfo Bongiovanni, Instituto Nacional de Tecnologia Agropecuaria (INTA), Manfredi, Cordoba, Argentina, (rodolfo_bongiovanni@yahoo.com)
}
\usage{
lasrosas2
}
\description{
Yield data and Nitrogen fertilizer treatment with field characteristics for the Las Rosas farm, Rio Cuarto, Cordoba, Argentina, 1999 (Rosas1999)
}
\details{
Sf object, undefined Coordinate Reference System.
}
\examples{
if (requireNamespace("sf", quietly = TRUE)) {
  library(sf)
  data(lasrosas2)

  plot(lasrosas2["YIELD"])
}
}
\keyword{datasets}
