% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liquorstore.R
\docType{data}
\name{liquorstore}
\alias{liquorstore}
\title{2015 Chicago Liquor Stores}
\format{
An sf data frame with 571 rows, 2 variables, and a geometry column:
\describe{
\item{	id	}{	Unique numeric ID	}
\item{	placeid	}{	Unique string ID	}
}
}
\source{
Scraped from Google Maps in 2015.
}
\usage{
liquorstore
}
\description{
Liquor stores in Chicago, IL.
}
\details{
Sf object, Transverse Mercator projection, GRS 1980.
}
\examples{
if (requireNamespace("sf", quietly = TRUE)) {
  library(sf)
  data(liquorstore)

  plot(liquorstore["id"])
}
}
\keyword{datasets}
