% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liquorstr.R
\docType{data}
\name{liquorstr}
\alias{liquorstr}
\title{2015 Chicago Liquor Stores}
\format{
An sf data frame with 571 rows, 3 variables, and a geometry column:
\describe{
\item{	id	}{	Unique numeric ID	}
\item{	Place ID	}{	Unique string ID	}
}
}
\source{
Scraped from Google Maps in 2015.
}
\usage{
liquorstr
}
\description{
Liquor stores in Chicago, IL.
}
\details{
Sf object, Transverse Mercator projection, GRS 1980.
}
\examples{
if (requireNamespace("sf", quietly = TRUE)) {
  library(sf)
  data(liquorstr)

  plot(liquorstr["id"])
}
}
\keyword{datasets}
