% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_col_dept.R
\docType{data}
\name{malaria_col_dept}
\alias{malaria_col_dept}
\title{Malaria in Colombia (1998).}
\format{
An sf data frame with 77 rows, 4 variables, and a geometry column:
\describe{
  \item{community}{Community name}
  \item{area_num_1}{Community ID}
  \item{NID}{Community ID (repeated)}
  \item{POP2010}{Population in 2010}
  \item{geometry}{MULTIPOLYGON}
}
}
\source{
\url{https://data.cityofchicago.org/Facilities-Geographic-Boundaries/Boundaries-Community-Areas-current-/cauq-8yn6}
}
\usage{
malaria_col_dept
}
\description{
Population in Chicago community areas in 2010. ### CHANGE IT !!!
}
\details{
Sf object, unprojected. EPSG 4326: WGS84.
}
\examples{
if (requireNamespace("sf", quietly = TRUE)) {
  library(sf)
  data(chicago_comm)

  plot(chicago_comm["community"])
}
}
\keyword{datasets}
