% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_col_munic.R
\docType{data}
\name{malaria_col_munic}
\alias{malaria_col_munic}
\title{Malaria in Colombia (1998)}
\format{
An sf data frame with 1068 rows, 51 variables, and a geometry column:
\describe{
	 \item{ID}{Sequencial ID}
	 \item{ADM0}{Country name}
 \item{ADM1}{Departments=States (Administrative level 1)}
 \item{ADM2}{Municipalities=Counties (Administrative level 2)}
 \item{IDDANE}{DANE (Colombian National Bureau of Statistics) code for municipalities in numeric format}
 \item{CODDEPT}{DANE (Colombian National Bureau of Statistics) code for municipalities}
 \item{TP1973}{Total population from census in 1973}
 \item{UP1973}{Urban population from census in 1973}
 \item{RP1973}{Rural population from census in 1973}
 \item{TP1985}{Total population from census in 1985}
 \item{UP1985}{Urban population from census in 1985}
 \item{RP1985}{Rural population from census in 1985}
 \item{TP1993}{Total population from census in 1993}
 \item{UP1993}{Urban population from census in 1993}
 \item{RP1993}{Rural population from census in 1993}
 \item{MALARI98}{Malaria incidence in 1998.}
 \item{Tpyyyy}{Total population projected in year yyyy}
 \item{Upyyyy}{Urban population projected in year yyyy}
 \item{Rpyyyy}{Rural population projected in year yyyy}
}
}
\source{
SIVIGILA. Ministerio de Salud, Colombia. There are 3 files for each year from 1995-2005. Projections from Departamento Administrativo Nacional de Estadistica DANE, Colombia \url{https://www.minsalud.gov.co} \url{https://www.dane.gov.co/index.php/estadisticas-por-tema/demografia-y-poblacion/proyecciones-de-poblacion}
}
\usage{
malaria_col_munic
}
\description{
Malaria incidence and population in Colombian municipalities from 1973, 95, 93 censuses and projections until 2005.
}
\details{
Sf object, unprojected. EPSG 4326: WGS84.
}
\examples{
if (requireNamespace("sf", quietly = TRUE)) {
  library(sf)
  data(malaria_col_munic)

  plot(malaria_col_munic["MALARI98"])
}
}
\keyword{datasets}
