% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncovr.R
\docType{data}
\name{ncovr}
\alias{ncovr}
\title{Homicides & Socio-Economics (1960-90)}
\format{
An sf data frame with 3,085 rows, 69 variables, and a geometry column:
\describe{
  \item{	NAME	}{	County name	}
  \item{	STATE_NAME	}{	State name	}
  \item{	STATE_FIPS	}{	State fips code (character)	}
  \item{	CNTY_FIPS	}{	County fips code (character)	}
  \item{	FIPS	}{	Combined state and county fips code (character)	}
  \item{	STFIPS	}{	State fips code (numeric)	}
  \item{	COFIPS	}{	County fips code (numeric)	}
  \item{	FIPSNO	}{	Fips code as numeric variable	}
  \item{	SOUTH	}{	Dummy variable for Southern counties  (South = 1)	}
  \item{	HRyy	}{	Homicide rate per 100,000 for the years yy: 1960, 1970, 1980 and 1990.	}
  \item{	HCyy	}{	Homicide count, three year average centered on yy: 1960, 1970, 1980, 1990	}
  \item{	POyy	}{	County population for the years yy: 1960, 1970, 1980, 1990	}
  \item{	RDyy	}{	Resource deprivation for the years yy: 1960, 1970, 1980, 1990 (principal component)	}
  \item{	PSyy	}{	Population structure for the years yy: 1960, 1970, 1980, 1990 (principal component)	}
  \item{	UEyy	}{	Unemployment rate for the years yy: 1960, 1970, 1980, 1990	}
  \item{	DVyy	}{	Divorce rate for the years yy: 1960, 1970, 1980, 1990  (percentage of males over 14 divorced)	}
  \item{	MAyy	}{	Median age for the years yy: 1960, 1970, 1980, 1990	}
  \item{	POLyy	}{	Log of population for the years yy: 1960, 1970, 1980, 1990	}
  \item{	DNLyy	}{	Log of population density for the years yy: 1960, 1970, 1980, 1990	}
  \item{	MFILyy	}{	Log of median family income for the years yy: 1960, 1970, 1980, 1990	}
  \item{	FPyy	}{	Percentage of families below poverty for the years yy: 1960, 1970, 1980, 1990 	}
  \item{	BLKyy	}{	Percentage of black for the years yy: 1960, 1970, 1980, 1990	}
  \item{	GIyy	}{	Gini index of family income inequality for the years yy: 1960, 1970, 1980, 1990	}
  \item{	FHyy	}{	Percentage of female headed households for the years yy: 1960, 1970, 1980, 1990	}
}
}
\source{
S. Messner, L. Anselin, D. Hawkins, G. Deane, S. Tolnay, R. Baller (2000). An Atlas of the Spatial Patterning of County-Level Homicide, 1960-1990. Pittsburgh, PA, National Consortium on Violence Research (NCOVR). \url{https://geodacenter.github.io/data-and-lab/ncovr/}
}
\usage{
ncovr
}
\description{
Homicides and selected socio-economic characteristics for continental U.S. counties. Data for four decennial census years: 1960, 1970, 1980 and 1990.
}
\details{
Sf object, unprojected. EPSG 4326: WGS84.
}
\examples{
if (requireNamespace("sf", quietly = TRUE)) {
  library(sf)
  data(ncovr)

  plot(ncovr["PO60"])
}
}
\keyword{datasets}
