% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyc.R
\docType{data}
\name{nyc}
\alias{nyc}
\title{Rental Housing and Demographics in NYC (2000s), non-spatial.}
\format{A data frame with 55 rows and 34 variables:
\describe{
  \item{FORHIS06-09}{percentage of hispanic population, not born in US}
  \item{FORWH06-09}{percentage of white population, not born in US}
  \item{HHSIZ1990}{average number of people per household}
  \item{HHSIZ00}{average number of people per household}
  \item{HHSIZ02-05-08}{average number of people per household}
  \item{KIDS2000, KIDS2005-2009}{percentage households w kids under 18}
  \item{RENT2002,2005,2008}{median monthly contract rent}
  \item{RENTPCT02,05,08}{percentage of housing stock that is market rate rental units}
  \item{PUBAST90,00}{percentage of households receiving public assistance}
  \item{YRHOMO02,05,08}{average number of years living in current residence}
}}
\source{
\url{https://geodacenter.github.io/data-and-lab/nyc/}
}
\usage{
nyc
}
\description{
Demographic and housing data for New York City’s 55 sub-boroughs (2000s).
}
\details{
Dataframe, no spatial components.
}
\keyword{datasets}
