% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nycearnings.R
\docType{data}
\name{nycearnings}
\alias{nycearnings}
\title{Block-level Earnings in NYC (2002-14)}
\format{
An sf data frame with 108,487 rows, 66 variables, and a geometry column:
\describe{
\item{	STATEFP10	}{	State FIPS code	}
\item{	COUNTYFP10	}{	County FIPS code	}
\item{	TRACTCE10	}{	Tract FIPS code	}
\item{	BLOCKCE10	}{	Block FIPS code	}
\item{	GEOID10	}{	Unique Geographic identifier (FIPS code)	}
\item{	UR10	}{	Urban (U) - rural (R) classification	}
\item{	UACE10	}{	Urban area census code 	}
\item{	UATYP10	}{	Urban type area. U: Urbanized areas of 50,000 people or more.	C: Urban Clusters (UCs) of at least 2,500 and less than 50,000 people.}
\item{	ALAND10	}{	Land area	}
\item{	AWATER10	}{	Water area	}
\item{	INTPTLAT10	}{	Latitude }
\item{	INTPTLON10	}{	Longitude	}
\item{	Shape_area	}{	Polygon area	}
\item{	Shape_len	}{	Polygon length	}
\item{	C000_YR	}{	Total number of jobs in year YR	}
\item{	CE01_YR	}{	Number of jobs with earnings $1250/month or less in year YR	}
\item{	CE02_YR	}{	Number of jobs with earnings $1251/month to $3333/month in year YR	}
\item{	CE03_YR	}{	Number of jobs with earnings greater than $3333/month in year YR	}
}
}
\source{
Longitudinal Employer-Household Dynamics (LEHD). \url{https://lehd.ces.census.gov/data/lodes/LODES7/LODESTechDoc7.0.pdf}
}
\usage{
nycearnings
}
\description{
Block-level Earnings New York City (2002-14) from Longitudinal Employer-Household Dynamics (LEHD).
}
\details{
Sf object, USA Contiguous Albers Equal Area Conic. ESRI Projection 102003. Geographic units: 2010 Census blocks.
}
\examples{
if (requireNamespace("sf", quietly = TRUE)) {
  library(sf)
  data(nycearnings)

  plot(nycearnings["CE01_02"])
}
}
\keyword{datasets}
