% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyceducation.R
\docType{data}
\name{nyceducation}
\alias{nyceducation}
\title{NYC Education (2000)}
\format{
An sf data frame with 2,216 rows, 56 variables, and a geometry column:
\describe{
\item{	poly_id	}{	Unique ID	}
\item{	youth_drop	}{	Percentage of population age 16-19 that has dropped out of high school	}
\item{	per_mnrty	}{	Percentage of the population that is non-white	}
\item{	hs_drop	}{	Percentage of population age over25 that dropped out of high school	}
\item{	col_degree	}{	Percentage of population over age 25 that obtained at least a bachelor’s degree	}
\item{	per_asian	}{	Percentage of Asian population	}
\item{	per_black	}{	Percentage of black population	}
\item{	per_white	}{	Percentage of white population	}
\item{	ctlabel	}{	Census tract label	}
\item{	borocode	}{	Borough code	}
\item{	boroname	}{	Borough name	}
\item{	boroct2000	}{	Census tract coding	}
\item{	ntacode	}{	Neighborhood tabulation area code	}
\item{	ntaname	}{	Neighborhood tabulation area name	}
\item{	puma	}{	Public use microarea code	}
\item{	shape_leng	}{	Length of polygon border	}
\item{	shape_area	}{	Area of polygon	}
\item{	np_ct	}{	Number of nonprofits	}
\item{	mean_inc	}{	Mean income	}
\item{	pop1619	}{	Population age 16-19	}
\item{	dropout	}{	High school dropouts, age 16-19	}
\item{	enrollhs	}{	Population enrolled in high school, age 16-19	}
\item{	per_prv_sc	}{	Percentage of all students Population enrolled in private school	}
\item{	per_pub_sc	}{	Percentage of all students Population enrolled in public school	}
\item{	over3	}{	Population over 3-year-old	}
\item{	notenroll	}{	Population over 3-year-old, not Population enrolled in school	}
\item{	over3enroll	}{	Population over 3-year-old, Population enrolled in school	}
\item{	pubsch	}{	Population enrolled in public school	}
\item{	pub_pk	}{	Population enrolled in public pre-k	}
\item{	pub_k8	}{	Population enrolled in public k-8	}
\item{	pub_hs	}{	Population enrolled in public high school	}
\item{	pub_col	}{	Population enrolled in public college	}
\item{	privsch	}{	Population enrolled in private school	}
\item{	priv_pk	}{	Population enrolled in private pre-k	}
\item{	priv_k8	}{	Population enrolled in private k-8	}
\item{	priv_hs	}{	Population enrolled in private high school	}
\item{	priv_col	}{	Population enrolled in private college	}
\item{	over25	}{	Population over 25 years	}
\item{	subhs	}{	Population over 25 years with less than high school degree	}
\item{	hs	}{	Population over 25 years with high school degree	}
\item{	somecol	}{	Population over 25 years with some college	}
\item{	college	}{	Population over 25 years with bachelor’s degree	}
\item{	master	}{	Population over 25 years with master’s degree	}
\item{	prof	}{	Population over 25 years with professional degree	}
\item{	phd	}{	Population over 25 years with phd	}
\item{	white	}{	Total white population	}
\item{	black	}{	Total black population	}
\item{	asian	}{	Total Asian population	}
\item{	sub18	}{	Total population under 18 years old	}
\item{	gender_par	}{	Gender parity, 1=parity, higher = more males, lower = more females	}
\item{	male	}{	Male population 	}
\item{	female	}{	Female population 	}
\item{	school_ct	}{	Number of schools 	}
\item{	popdens	}{	Population density (people per square mile)	}
\item{	population	}{	Population count	}
}
}
\source{
Source for school locations: Open Data New York (\url{https://data.cityofnewyork.us/Education/School-Point-Locations/jfju-ynrr}). Demographics source: 2000 Census.
}
\usage{
nyceducation
}
\description{
Block-level New York City information about education and demographic characteristics (2000). Geographic units: 2010 Census blocks.
}
\details{
Sf object, Projection 2203. NAD83 / New York Long Island (ftUS)
}
\examples{
if (requireNamespace("sf", quietly = TRUE)) {
  library(sf)
  data(nyceducation)

  plot(nyceducation["phd"])
}
}
\keyword{datasets}
