% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozone96.R
\docType{data}
\name{ozone96}
\alias{ozone96}
\title{Ozone data 1996}
\format{
An sf data frame with 30 rows, 79 variables, and a geometry column:
\describe{
\item{	STATION	}{	monitoring station identifier	}
\item{	MAXDAY	}{	maximum daily ozone for the month	}
\item{	AV8TOP	}{	monthly average of highest 8 hour daily averages	}
\item{	MONITOR	}{	monitoring station identifier	}
\item{	LAT	}{	latitude	}
\item{	LON	}{	longitude	}
\item{	X_COORD	}{	X-coordinate, projected, UTM	}
\item{	Y_COORD	}{	Y-coordinate, projected, UTM	}
}
}
\source{
Original data provided by James Murdoch, School of Social Sciences, University of Texas at Dallas, Richardson, TX 75083 (murdoch@utdallas.edu).
}
\usage{
ozone96
}
\description{
July 1996 ozone measurements for Los Angeles basin monitoring stations.
}
\details{
Sf object, unprojected. EPSG 4326: WGS84.
}
\examples{
if (requireNamespace("sf", quietly = TRUE)) {
  library(sf)
  data(ozone96)

  plot(ozone96["M971"])
}
}
\keyword{datasets}
