% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIDS2.R
\docType{data}
\name{SIDS2}
\alias{SIDS2}
\title{Sudden Infant Death Syndrome (SIDS2)}
\format{
An sf data frame with 100 rows, 18 variables, and a geometry column:
\describe{
\item{	AREA	}{	county area (computed by ArcView)	}
\item{	PERIMETER	}{	county perimeter (computed by ArcView)	}
\item{	CNTY_	}{	county internal ID	}
\item{	CNTY_ID	}{	county internal ID	}
\item{	NAME	}{	county name	}
\item{	FIPS	}{	county fips code, as character (state code and county code)	}
\item{	FIPSNO	}{	county fips code, numeric, used in GeoDa User’s Guide and tutorials	}
\item{	CRESS_ID	}{	county ID used by Cressie	}
\item{	BIR74	}{	live births, 1974-78	}
\item{	SID74	}{	SIDS deaths, 1974-78	}
\item{	NWBIR74	}{	non-white births, 1974-78	}
\item{	BIR79	}{	live births, 1979-84	}
\item{	SID79	}{	SIDS deaths, 1979-84	}
\item{	NWBIR79	}{	non-white births, 1979-84	}
\item{	SIDR74	}{	SIDS death rate per 1,000 (1974-78)	}
\item{	SIDR79	}{	SIDS death rate per 1,000 (1979-84)	}
\item{	NWR74	}{	non-white birth rate (non-white per 1000 births), 1974-78	}
\item{	NWR79	}{	non-white birth rate (non-white per 1000 births), 1979-84	}
}
}
\source{
Cressie, Noel (1993). Statistics for Spatial Data. New York, Wiley, pp. 386-389. Rates computed.
}
\usage{
SIDS2
}
\description{
Sudden Infant Death Syndrome sample data for North Carolina counties, two time periods (1974-78 and 1979-84). Same as SIDS data set, except that the computed rates are included.
}
\details{
Sf object, unprojected. EPSG 4326: WGS84.
}
\examples{
if (requireNamespace("sf", quietly = TRUE)) {
  library(sf)
  data(SIDS2)
  plot(SIDS2["BIR74"])
}
}
\keyword{datasets}
