% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_sdoh.R
\docType{data}
\name{us_sdoh}
\alias{us_sdoh}
\title{US Social Determinants of Health Data 2014}
\format{

}
\source{
\url{https://sdohatlas.github.io/}
}
\usage{
us_sdoh
}
\description{
Social Determinants of Health Data 2014 from the United States at neighborhood-level
}
\details{
Sf object. EPSG 4326: WGS84.
}
\examples{
if (requireNamespace("sf", quietly = TRUE)) {
  library(sf)
  data(us_sdoh)

  plot(us_sdoh["ep_pov"])
}
}
\keyword{datasets}
