% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chigroceries.R
\docType{data}
\name{chigroceries}
\alias{chigroceries}
\title{2015 Chicago supermarkets}
\format{
An sf data frame with 148 rows, 7 variables, and a geometry column:
\describe{
\item{	OBJECTID	}{	Location ID	}
\item{	Xcoord	}{	X-coordinate	}
\item{	Ycoord	}{	Y-coordinate	}
\item{	Status	}{	Status of the store (e.g. open, open at new location, etc.)	}
\item{	Address	}{	Supermarket address	}
\item{	Chain	}{	Supermarket chain	}
\item{	Category	}{	New stores are marked “New”, the rest are left blank	}
}
}
\source{
Scraped from Google Maps in 2015.
}
\usage{
chigroceries
}
\description{
Grocery stores in Chicago, IL as of 2015.
}
\details{
Sf object, Transverse Mercator projection: GRS 80.
}
\examples{
if (requireNamespace("sf", quietly = TRUE)) {
  library(sf)
  data(chigroceries)

  plot(chigroceries["Ycoord"])
}
}
\keyword{datasets}
