% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodaData-package.R
\docType{package}
\name{geodaData-package}
\alias{geodaData}
\alias{geodaData-package}
\title{geodaData: Spatial Analysis Datasets for Teaching}
\description{
The goal of geodaData is to store sample spatial datasets. These 
    datasets are intended to be used to teach basic spatial analysis concepts. 
    They are used in the weekly R Spatial Workshop at the Center for Spatial Data 
    Science at UChicago, and are based off of the GeoDa workbook and data site 
    developed by Luc Anselin and team. Datasets are stored in the sf object format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/spatialanalysis/geodaData}
  \item Report bugs at \url{https://github.com/spatialanalysis/geodaData/issues}
}

}
\author{
\strong{Maintainer}: Angela Li \email{ali6@uchicago.edu} (\href{https://orcid.org/0000-0002-8956-419X}{ORCID})

}
\keyword{internal}
