% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nycearnings.R
\docType{data}
\name{nycearnings}
\alias{nycearnings}
\title{Block-level Earnings in NYC (2002-14)}
\format{
An sf data frame with 108487 rows, 70 variables, and a geometry column:
\describe{
\item{	YR	}{	02 to 14 (2002 to 2014)	}
\item{	GEOID10	}{	unique ID (FIPS code)	}
\item{	C000_YR	}{	Total number of jobs	}
\item{	CE01_YR	}{	Number of jobs with earnings $1250/month or less in year YR	}
\item{	CE02_YR	}{	Number of jobs with earnings $1251/month to $3333/month in year YR	}
\item{	CE03_YR	}{	Number of jobs with earnings greater than $3333/month in year YR	}
}
}
\source{
Longitudinal Employer-Household Dynamics (LEHD). \url{https://lehd.ces.census.gov/data/lodes/LODES7/LODESTechDoc7.0.pdf}
}
\usage{
nycearnings
}
\description{
Block-level Earnings New York City (2002-14) from Longitudinal Employer-Household Dynamics (LEHD).
}
\details{
Sf object, USA Contiguous Albers Equal Area Conic. ESRI Projection 102003. Geographic units: 2010 Census blocks.
}
\examples{
if (requireNamespace("sf", quietly = TRUE)) {
  library(sf)
  data(nycearnings)

  plot(nycearnings["CE01_02"])
}
}
\keyword{datasets}
