% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/southom.R
\docType{data}
\name{southom}
\alias{southom}
\title{Homicides in the South (1960-1990)}
\format{
An sf data frame with 1412 rows, 70 variables, and a geometry column:
\describe{
\item{	NAME	}{	county name	}
\item{	STATE_NAME	}{	state name	}
\item{	STATE_FIPS	}{	state fips code (character)	}
\item{	CNTY_FIPS	}{	county fips code (character)	}
\item{	FIPS	}{	combined state and county fips code (character)	}
\item{	STFIPS	}{	state fips code (numeric)	}
\item{	COFIPS	}{	county fips code (numeric)	}
\item{	FIPSNO	}{	fips code as numeric variable	}
\item{	SOUTH	}{	dummy variable for Southern counties (South = 1)	}
\item{	HR**	}{	homicide rate per 100,000 (1960, 1970, 1980, 1990)	}
\item{	HC**	}{	homicide count, three year average centered on 1960, 1970, 1980, 1990	}
\item{	PO**	}{	county population, 1960, 1970, 1980, 1990	}
\item{	RD**	}{	resource deprivation 1960, 1970, 1980, 1990 (principal component, see Codebook for details)	}
\item{	PS**	}{	population structure 1960, 1970, 1980, 1990 (principal component, see Codebook for details)	}
\item{	UE**	}{	unemployment rate 1960, 1970, 1980, 1990	}
\item{	DV**	}{	divorce rate 1960, 1970, 1980, 1990 (pct of males over 14 divorced)	}
\item{	MA**	}{	median age 1960, 1970, 1980, 1990	}
\item{	POL**	}{	log of population 1960, 1970, 1980, 1990	}
\item{	DNL**	}{	log of population density 1960, 1970, 1980, 1990	}
\item{	MFIL**	}{	log of median family income 1960, 1970, 1980, 1990	}
\item{	FP**	}{	pct families below poverty 1960, 1970, 1980, 1990 (see Codebook for details)	}
\item{	BLK**	}{	pct black 1960, 1970, 1980, 1990	}
\item{	GI**	}{	Gini index of family income inequality 1960, 1970, 1980, 1990	}
\item{	FH**	}{	pct female headed households 1960, 1970, 1980, 1990	}
}
}
\source{
S. Messner, L. Anselin, D. Hawkins, G. Deane, S. Tolnay, R. Baller (2000). An Atlas of the Spatial Patterning of County-Level Homicide, 1960-1990. Pittsburgh, PA, National Consortium on Violence Research (NCOVR) (\url{www.ncovr.heinz.cmu.edu}).
}
\usage{
southom
}
\description{
Homicides and selected socio-economic characteristics for Southern U.S. counties (subset of NCOVR national data set). Data for four decennial census years: 1960, 1970, 1980, 1990.
}
\details{
Sf object, unprojected. EPSG 4326: WGS84.
}
\examples{
if (requireNamespace("sf", quietly = TRUE)) {
  library(sf)
  data(southom)
  plot(southom["HR90"])
}
}
\keyword{datasets}
